Removed default Rake task
Temporarily disabled in order to allow build to pass since there are
commits on `master` that are not in the right format.

It looks like Netlify has changed how they clone projects when building
so Git Cop is picking up commits on `master` instead of the feature
branchs when linting.